feat: use ofetch imitate got
