+++ package.json
@@ -63,6 +63,7 @@
         "crypto-js": "4.2.0",
         "currency-symbol-map": "5.1.0",
         "dayjs": "1.11.8",
+        "destr": "2.0.3",
         "directory-import": "3.2.1",
         "dotenv": "16.4.5",
         "entities": "4.5.0",
@@ -89,6 +90,7 @@
         "module-alias": "2.2.3",
         "notion-to-md": "3.1.1",
         "oauth-1.0a": "2.2.6",
+        "ofetch": "1.3.4",
         "otplib": "12.0.1",
         "pac-proxy-agent": "7.0.1",
         "proxy-chain": "2.4.0",
